% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{rle_regions}
\alias{rle_regions}
\title{Contiguous TRUE regions to start/end data.frame}
\usage{
rle_regions(above, trial_idx)
}
\description{
Contiguous TRUE regions to start/end data.frame
}
\keyword{internal}
